requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0

# Database & API Client
supabase==2.3.7
httpx==0.26.0

# Security & Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6

# Configuration & Validation
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Email Service
# standard smtplib is included in Python, but these help with advanced formatting
jinja2==3.1.3